+++ backend/canisters/local_user_index/impl/src/updates/c2c_notify_user_index_events.rs
@@ -3,6 +3,7 @@ use crate::{CanisterToRefund, CommunityEvent, GroupEvent, RuntimeState, UserEven
 use canister_api_macros::update;
 use canister_time::now_millis;
 use canister_tracing_macros::trace;
+use group_canister::UserIdMigrated as GroupUserIdMigrated;
 use local_user_index_canister::c2c_notify_user_index_events::*;
 use local_user_index_canister::{UserIndexEvent, UserRegistered};
 use p256_key_pair::P256KeyPair;
@@ -359,7 +360,20 @@ fn handle_event<F: FnOnce() -> TimestampMillis>(
             state.set_daily_puzzle_canister_id(canister_id);
         }
         UserIndexEvent::UserIdMigrated(ev) => {
-            state.data.migrated_user_ids.insert(ev.old_user_id, ev.new_user_id);
+            if state.data.migrated_user_ids.insert(ev.old_user_id, ev.new_user_id) {
+                for chat_id in ev.groups {
+                    if state.data.local_groups.get(&chat_id).is_some() {
+                        state.push_event_to_group(
+                            chat_id.into(),
+                            GroupEvent::UserIdMigrated(GroupUserIdMigrated {
+                                old_user_id: ev.old_user_id,
+                                new_user_id: ev.new_user_id,
+                            }),
+                            **now,
+                        );
+                    }
+                }
+            }
         }
     }
 }