+++ backend/canisters/user_index/impl/src/lifecycle/inspect_message.rs
@@ -66,6 +66,7 @@ fn accept_if_valid(state: &RuntimeState) {
         | "upgrade_multi_user_canister_wasm"
         | "upgrade_local_user_index_canister_wasm"
         | "mark_local_user_index_full"
+        | "record_user_id_migrated"
         | "register_external_achievement"
         | "publish_bot"
         | "suspected_bots" => state.is_caller_governance_principal(),